Trying to fix the llr and gvt.NPV calculations
lots of updates to model status.
futzing around with modelinit and bank_hurdle.R to try to fix some
things, here are my notes on debugging them so far:

Why are calculated user interest rates insensitive to risk of default?
Now (running line by line) it appears that they are not! Interest.rate
changes from the bank.hurdle, but the interest rate we’re putting in the
output matrix is different (though it shouldn’t be)
Loan.payment appears to be the same as loan.payment.user
Gvt.cost.NPV is almost zero when running line by line from bank_hurdle…
even though it isn’t in the results output!!!
SOMEHOW I CHANGED SOOMETHING THAT FIXED EVERYTHING WITHOUT KNOWING IT!
Just re-ran modelinit and it gave an almost (BUT NOT QUITE) zero cost of
government and a different interest.user from the bank.hurdle. Although
it looks like the loan.payment and loan.payment.user are different for
the chance.full.loss=.99 scenario…