Functional model + do-dos
Fixed bank.NPV, so that it is zero, unless there is an IRB and a nonzero
chance of default.

added a version of modelinit that is made with the inputs from the
zimring paper/DOE model

Commented bank.hurdle with notes from 3/27 on upcoming additions to the
model

Model_Status has some updates from that meeting, too.